[WIP] Correction for the controlled systems
Rewrites the cases (forced, constrained).
Separates the examples.
Rewrites the controlled systems, addition of a new example (pendulum controlled by the response of optimal control) .
Needs to be read carefully.
