+++ biorbd_examples/pendulum_control.py
@@ -1,0 +1,83 @@
+"""
+This script is used to integrate the motion with a variational integrator based on the discrete Lagrangian,
+and a first order quadrature method.
+"""
+import biorbd_casadi
+
+from varint.minimal_variational_integrator import VariationalIntegrator
+
+from utils import *
+
+
+def pendulum():
+    biorbd_casadi_model = biorbd_casadi.Model(Models.PENDULUM.value)
+    biorbd_model = biorbd.Model(Models.PENDULUM.value)
+
+    import time as t
+
+    time = 1
+    time_step = 1/30
+
+    tic0 = t.time()
+
+    tau = np.asarray([[
+        18.39841936, 18.56183951,  18.52361453,  18.28517695,  17.85120277,
+        17.2300812,  16.43428248,  15.48044982,  14.38906786,  13.18363429,
+        11.88936988, 10.5316134,   9.13412692,   7.71755652,   6.29825044,
+        4.88754928,  3.49156338,   2.1113697,    0.74351136,   -0.61932967,
+        -1.98760312, -3.37419226,  -4.79380495,  -6.26248594,  -7.79726463,
+        -9.41594264, -11.13702162, -12.97977061, -14.96443099, -17.11254998
+    ]])
+
+    # variational integrator
+    vi = VariationalIntegrator(
+        biorbd_model=biorbd_casadi_model,
+        time_step=time_step,
+        time=time,
+        q_init=np.array([[0.0, 0.01030821]]),
+        controls=tau,
+    )
+
+    q_vi, _ = vi.integrate()
+
+    print(q_vi)
+
+    q_optimal_control = [
+        0.        , 0.01030821, 0.04121574, 0.09245982, 0.16332607,
+        0.25266637, 0.35893181, 0.48022312, 0.61435943, 0.75896352,
+        0.91155867, 1.06966929, 1.23091611, 1.39309694, 1.55424584,
+        1.712667  , 1.86694296, 2.01591981, 2.15867424, 2.29446765,
+        2.42269275, 2.54281691, 2.65432548, 2.75666724, 2.8492029 ,
+        2.93115699, 3.00157283, 3.05926975, 3.10280139, 3.1304136 ,
+        3.14
+    ]
+
+    tic1 = t.time()
+    print(tic1 - tic0)
+
+    import bioviz
+
+    b = bioviz.Viz(Models.PENDULUM.value)
+    b.load_movement(q_vi)
+    b.exec()
+
+    import matplotlib.pyplot as plt
+
+    plt.figure()
+    plt.plot(q_vi[0, 1:], label="Variational Integrator", color="red", linestyle="-", marker="", markersize=2)
+    plt.title(f"Generalized coordinates")
+    plt.legend()
+
+    # Plot total energy for both methods
+    plt.figure()
+    plt.plot(discrete_total_energy(biorbd_model, q_vi, time_step), label="Variational Integrator", color="red")
+    plt.title(f"Total energy")
+    plt.legend()
+
+    plt.show()
+
+    return print("Hello World")
+
+
+if __name__ == "__main__":
+    pendulum()
